pyproject.toml
[tool.poetry]
name = "micronet"
version = "0.1.0"
description = "A package to facilitate network simulations"
authors = ["Christian Hopps <chopps@labn.net>"]
license = "GPL-2.0-or-later"
readme = "README.md"
repository = "https://github.com/LabNConsulting/micronet"

[tool.poetry.dependencies]
python = "^3.8"
toml = {version = "^0.10.2", optional = true}
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pylint = "^2.11.1"
pytest = "^6.2.5"

[tool.poetry.scripts]
micronet = "micronet.__main__:main"

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["micronet", "tests"]
skip_glob = ["*/setup.py"]
filter_files = true
known_first_party = "micronet"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
ignore = "E203"

[tool.pytest.ini_options]

testpaths = [
        "tests",
]

log_level = "DEBUG"
log_format = "%(asctime)s,%(msecs)03d %(levelname)s: %(name)s: %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

# If verbose is specifyied log_cli will be set to 1, it can also be specified
# here or on the CLI.
# log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s,%(msecs)03d %(levelname)s: %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

log_file_level = "DEBUG"
log_file_format = "%(asctime)s,%(msecs)03d %(levelname)s: %(name)s: %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
